requirements.txt
environs==11.0.0
requests==2.31.0
python-telegram-bot==21.2